+++ tpu-client/src/tpu_client.rs
@@ -28,8 +28,7 @@ use {
 pub const DEFAULT_VOTE_USE_QUIC: bool = false;
 
 /// The default connection count is set to 1 -- it should
-/// be sufficient for most use cases. Validators can use
-/// --tpu-connection-pool-size to override this default value.
+/// be sufficient for most use cases.
 pub const DEFAULT_TPU_CONNECTION_POOL_SIZE: usize = 1;
 
 pub type Result<T> = std::result::Result<T, TpuSenderError>;